+++ src/main/java/com/interview/bankApp/repository/TransactionRepository.java
@@ -3,5 +3,5 @@ package com.interview.bankApp.repository;
 import com.interview.bankApp.model.Transaction;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TransactionRepository extends CrudRepository<Transaction, Integer> {
+public interface TransactionRepository extends CrudRepository<Transaction, Long> {
 }
